[DD-902] refactored/moved some classes to model package (from delta)
