+++ src/main/java/org/dswarm/graph/delta/util/GraphDBUtil.java
@@ -30,9 +30,9 @@ import java.util.Set;
 import org.dswarm.graph.DMPGraphException;
 import org.dswarm.graph.GraphIndexStatics;
 import org.dswarm.graph.NodeType;
-import org.dswarm.graph.delta.Attribute;
-import org.dswarm.graph.delta.AttributePath;
-import org.dswarm.graph.delta.ContentSchema;
+import org.dswarm.graph.model.Attribute;
+import org.dswarm.graph.model.AttributePath;
+import org.dswarm.graph.model.ContentSchema;
 import org.dswarm.graph.delta.DMPStatics;
 import org.dswarm.graph.delta.DeltaState;
 import org.dswarm.graph.delta.DeltaStatics;
